refactor : 2단계 리뷰사항 반영

+++ src/main/java/nextstep/member/application/MemberService.java
@@ -1,12 +1,14 @@
 package nextstep.member.application;
 
 import nextstep.exception.NotFoundUserException;
+import nextstep.member.AuthenticationException;
 import nextstep.member.application.request.MemberRequest;
 import nextstep.member.application.response.MemberResponse;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class MemberService {
@@ -16,22 +18,35 @@ public class MemberService {
         this.memberRepository = memberRepository;
     }
 
+    @Transactional
     public MemberResponse createMember(MemberRequest request) {
         Member member = memberRepository.save(Member.of(request.getEmail(), request.getPassword(), request.getAge()));
         return MemberResponse.of(member);
     }
 
-    public MemberResponse findMember(Long id) {
+    public MemberResponse findMember(LoginMember loginMember, Long id) {
         Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (!member.getMemberId().equals(id) && !loginMember.getEmail().equals(member.getEmail())) {
+            throw new AuthenticationException();
+        }
         return MemberResponse.of(member);
     }
 
-    public void updateMember(Long id, MemberRequest param) {
+    @Transactional
+    public void updateMember(LoginMember loginMember, Long id, MemberRequest param) {
         Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (!member.getMemberId().equals(id) && !loginMember.getEmail().equals(member.getEmail())) {
+            throw new AuthenticationException();
+        }
         member.update(param.toMember());
     }
 
-    public void deleteMember(Long id) {
+    @Transactional
+    public void deleteMember(LoginMember loginMember, Long id) {
+        Member member = findMemberByEmail(loginMember.getEmail());
+        if (!member.getMemberId().equals(id)) {
+            throw new AuthenticationException();
+        }
         memberRepository.deleteById(id);
     }
 
@@ -44,4 +59,5 @@ public class MemberService {
                 .map(it -> MemberResponse.of(it))
                 .orElseThrow(RuntimeException::new);
     }
+
 }
